save constellation to playlist image
